+++ dags/brz_industry_code_monthly/constants.py
