+++ pyproject.toml
@@ -1,0 +1,4 @@
+[tool.ruff]
+lint.select = ["I"]
+fix = true
+line-length = 88
